Extract stringplus package
