+++ exo-go/src/stringplus/join_string_maps.go
@@ -1,4 +1,4 @@
-package util
+package stringplus
 
 // JoinStringMaps joins the given string maps
 func JoinStringMaps(maps ...map[string]string) map[string]string {
