+++ exo-go/src/stringplus/join_string_slices.go
@@ -1,4 +1,4 @@
-package util
+package stringplus
 
 // JoinStringSlices joins the given slices
 func JoinStringSlices(slices ...[]string) []string {
